requirements.txt
PyQt5
requests
pytest
pytest-qt
flake8
black
python-dotenv
